+++ test/test_dbcore.py
@@ -24,9 +24,9 @@ from typing import ClassVar
 import pytest
 
 from beets import dbcore
-from beets.dbcore import query, sort
+from beets.dbcore import query, sort, types
 from beets.dbcore.db import DBCustomFunctionError, Index
-from beets.library import LibModel
+from beets.library import Album, Item, LibModel
 from beets.util import cached_classproperty
 
 # Fixture: concrete database and model classes. For migration tests, we
@@ -538,6 +538,26 @@ class ParseTest(unittest.TestCase):
     def test_parse_untyped_field(self):
         value = ModelFixture1._parse("field_nine", "2")
         assert value == "2"
+
+
+class TestModelTypeFallback:
+    def test_album_type_falls_back_to_item_type(self):
+        typ = Album._type("artists")
+        assert isinstance(typ, types.DelimitedString)
+        assert typ is types.MULTI_VALUE_DSV
+
+    def test_album_type_falls_back_to_item_type_other_list_fields(self):
+        for field in ["genres", "composers", "artists_sort"]:
+            typ = Album._type(field)
+            assert isinstance(typ, types.DelimitedString), field
+
+    def test_item_type_does_not_change(self):
+        typ = Item._type("artists")
+        assert isinstance(typ, types.DelimitedString)
+
+    def test_unknown_key_falls_through_to_default(self):
+        typ = Album._type("nonexistent_field_xyz")
+        assert isinstance(typ, types.Default)
 
 
 class QueryParseTest(unittest.TestCase):